Add TemplateReactionWithReverse(TemplateReaction)
Created a subclass of TemplateReaction, called
TemplateReactionWithReverse, specifically for templateReactions generated
by the families which are ownReverse. Previously such templateReaction has
attribute "reverse" which is neither defined in __init__ of
TemplateReaction nor in __reduce__ method. So they have problems pickling
when running in parallel mode. With this approach, all the reactions can
be pickled successfully.