Isolated study of react_parallel
In order just to test the time saving for react part, it's always better
not to run RMG job but only run react (or react_species,
react_speciesList, react_family) and compare it directly with serial mode.
So this commit layout such an environment.
1. isolated_react_test.py is the testing script, currently in case1 mode,
which doesn't use IDize for species, use family objects and template
objects in reactions to serilize.
2. family.py is also in case1 scenario.
3. model.py is in case3 scenario, which use IDize and family labels and
template labels to reduce communication cost. In addition, react_species
and react_speciesList added timing recording for getConst and genRxn.